Python requires consistent indentation.
The AdamOptimizer() function's indentation has some lines that are all
spaces and some are mixed spaces and tabs. At least with Python 3.6.8,
this causes a

TabError: inconsistent use of tabs and spaces in indentation

Switched all indentation in the function to be tab-based, since the
rest of the file is tab indented.